added images

+++ src/core/Files.php
@@ -11,7 +11,8 @@ class Files
         $storedFiles = [];
 
         foreach ($files as $file) {
-            $target_file = Application::$ROOT_DIR . "/storage/" . uniqid("f_");
+            $name  = uniqid("f_");
+            $target_file = Application::$ROOT_DIR . "/storage/" . $name;
             $ext = ".". strtolower(pathinfo($file["name"],PATHINFO_EXTENSION));
 
             $uploaded = move_uploaded_file($file["tmp_name"], $target_file . $ext);
@@ -19,7 +20,7 @@ class Files
             if ($uploaded) {
                 $storedFiles[] = File::create([
                     "name" => $file["name"],
-                    "path" => $target_file,
+                    "path" => "/storage/" . $name,
                     "ext" => $ext,
                 ]);
             }
@@ -31,6 +32,11 @@ class Files
     public static function delete($id)
     {
         $file = File::where(["id", $id]);
-        return $file->delete();
+        if ($file) {
+            unlink(Application::$ROOT_DIR . $file->path . $file->ext);
+            return $file->delete();
+        }
+
+        return false;
     }
 }
